remove old code + tests
