--- tests/Rules/Validators/BooleanTest.php
@@ -1,60 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Gocanto\Attributes\Tests\Rules\Validators;
-
-use Gocanto\Attributes\AttributesException;
-use Gocanto\Attributes\Rules\Constraint;
-use Gocanto\Attributes\Rules\Validators\Boolean;
-use PHPUnit\Framework\TestCase;
-
-class BooleanTest extends TestCase
-{
-    /**
-     * @test
-     */
-    public function itHoldsValidData()
-    {
-        $constraint = new Boolean();
-
-        $this->assertInstanceOf(Constraint::class, $constraint);
-        $this->assertSame('boolean', $constraint->getIdentifier());
-    }
-
-    /**
-     * @test
-     * @throws AttributesException
-     */
-    public function itProperlyValidateTheInputAndReturnGenericMessages()
-    {
-        $this->expectException(AttributesException::class);
-        $this->expectExceptionMessageMatches('/Expected a boolean. Got/');
-
-        $constraint = new Boolean();
-        $constraint->assert(1);
-    }
-
-    /**
-     * @test
-     * @throws AttributesException
-     */
-    public function itProperlyValidateTheInputAndReturnCustomMessages()
-    {
-        $this->expectException(AttributesException::class);
-        $this->expectExceptionMessageMatches('/Message error stub/');
-
-        $constraint = new Boolean();
-        $constraint->assert('foo', 'Message error stub');
-    }
-
-    /**
-     * @test
-     * @throws AttributesException
-     */
-    public function itAllowsValidValues()
-    {
-        $this->doesNotPerformAssertions();
-
-        $constraint = new Boolean();
-        $constraint->assert(true);
-    }
-}
